JIT: introduce durable EH region ID (#113497)
Give each EH clause a permanent ID unique to that clause. Use this
to revise how we model GT_END_LFIN's nesting depth -- the node now
refers to a region ID, and we use that plus a lookup table built
during `FinalizeEH` to figure out the nesting depth.

Contributes to #108900.