Fix x86 exception handling edge case
When a NULL reference exception occurs in a JIT helper or a VSD stub,
runtime pretends the exception occured in the managed caller. There is a
bug on x86 Windows where the COMPlusThrowCallback considers that frame
to be the frame where the exception actually occurred (based on the
m_crawl.isFirst). In case the call to the helper is the last instruction
in a try region, the exception handler lookup would reject that address
and the exception may not get handled at the right place or at all.

This change fixes it by ensuring that the m_crawl.isFirst is not set
when the frame is not the frame of the failure.

Close #113106