Add ThisptrRetbufPrecode to set of precodes handled by StubPrecode path (#112548)
This adds the ThisPtrRetBufPrecode to the set of codes handled by the StubPrecode stub pasting path.

Perf changes from about 2.35 ns per iteration of a minimal loop with a ThisPtrRetBufPrecode to about 3.03 ns per iteration on my computer (Windows X64). I expect the cost to be a little less on an Arm64 machine since the old ThisPtrRetBufPrecode was a bit more expensive there. However, this removes the previous scheme where we allocate and jit custom stubs for this, so I think the simplification is worth the runtime cost.

The overall change has a size increase though, as this work adds support for this type of precode to the cDAC, as well as some tests.