Modernize `CookieContainer` domain-matching (#112604)
With #39781 and #64038 we shifted the .NET `Cookie`/`CookieContainer` behavior towards RFC 6265 for handling the leading dot in the `Domain` attribute, however there were gaps left primarily because the the cookie code is complicated and full of logic that handles legacy requirements from old RFC-s related to now obsolete attributes.  The existing obscure logic results in weird unintended behaviors under the hood and the bug #84820.

This PR is addressing this problem by adapting the [RFC 6265 domain matching algorithm](https://datatracker.ietf.org/doc/html/rfc6265#section-5.1.3) with a few modifications in order to avoid breaking changes:
- Single-label domains: in order for a host string to match a single-label domain string, they have to be identical.
- `GetCookies(uri)` does not apply the RFC 6265 domain-matching for implicit domains. An implicit domain string has to be identical to host in order to match it.