Fix dynamic static generic virtual method dispatch on generic types (#113684)
Fixes #113664.

When we resolve dispatch for a static generic virtual method dynamically (e.g. after `MakeGenericType` or another generic virtual method call) we run the runtime type loader logic to do the resolution. The resolution was incorrectly returning an open version of the containing type and that was resulting in a failed dispatch. The version of the resolution algorithm in the compiler (handling the non-dynamic cases) didn't have the bug.

This behavior dates back to .NET Native times. It was never a problem in .NET Native because for instance generic virtual methods, we paper over the open version of the type as we resolve the slot on the implementing class (the interface method resolution happens in two phases: first we resolve the interface method to a virtual slot on the implementing class, and then we resolve the virtual slot to an actual implementation). It is a problem for static virtual methods since the second phase doesn't happen (the static virtual implementations need to be `sealed` in IL so there's no point to do the second pass).