JIT: revise escape analysis bit vector usage
This change decouples the "bv" space (used for the connection graph
and related data) from the "lclnum" space, so that the graph nodes can
represent other entities that are not locals.

Conditional escape analysis introduced pseudo-locals, so we already had
this notion, but it was implicit, and extending it further (for say
fields) was proving awkward. So now it is explicit.

The one tricky aspect is that as the analysis proceeds we may introduce
new locals that we want to track, so we need to anticipate this and
leave bv space for them, since our bv length must be known in advance.

Only a subset of locals need to be analyzed, so in some (many?) cases
the overall BV lengths should decrease, though there is now a bit more
work to interpret what the bits mean.

We reuse the tracked var concept and data structures for locals, eg to
map from bv index to local num; for other entities the mappings are
one-to-one so reverse mapping doesn't require table lookups.