JIT: Avoid rare self-edge in register homing
Only the secret stub parameter goes through this path to create an edge
in the graph, and it was missing a check for whether it was going into
the same register as it was in. This would cause register homing to
generate unnecessary moves into a temporary register because it thought
there was a conflict.

Usually the secret stub parameter has to be allocated to a callee saved
variable, so we typically do not see this inefficiency.