Change CORJIT_FLAG_READYTORUN and CORJIT_FLAG_PREJIT to CORJIT_FLAG_AOT (#113850)
CORJIT_FLAG_READYTORUN and CORJIT_FLAG_PREJIT were always both set
together. Introduce a new name to be more explicit of a concept
that includes both ReadyToRun and NativeAOT (AOT == Ahead Of Time).

Introduce new JIT helpers `IsAot` (for either NativeAOT or R2R),
`IsNativeAot`, and `IsReadyToRun`.

Remove the name "ngen" in a bunch of places in the JIT code.