Add support for fields
Loading and storing into instance and static fields, as well as loading the address of fields. For instance fields, this is achieved via ldind/stind opcodes, that have the offset embedded into the instruction. For static fields, we first compute the field address and then we reuse the same ldind/stind opcodes for the field access.

When loading a field from a valuetype from the interpreter stack, we use a separate opcode INTOP_MOV_SRC_OFF, which will be emitted as a normal move (with already existing opcodes), but the source offset that it operates on can point inside a VT var.