Stackwalk support for interpreted frames
This change adds support for walking stack with interpreter frames
present. The StackFrameIterator now considers the interpreter frames to
be frameless frames like regular JITted / crossgened folder.
This is a basis for GC, EH and debugger stack walking. The SOS
`clrstack` and `clrstack -i` both display correct stack trace with
interpreter frames on the stack with this change.

I have added some extra handling of the case when the interpreter frames
are on top of the stack so that the clrstack can dump them.