Fix AltJit and VectorT ISA specification (#113901)
* Fix AltJit and VectorT ISA specification

When running as an AltJit when the generated code will not be used
(either because of a mismatched VM, or because RunAltJitCode=0),
the JIT overwrites the passed in ISAs. When it does so, it zeros
out the VectorT* ISA flags.

Instead of this, carry over the existing ISA flags. This should work
for same-architecture SPMI replays.

I'm not sure if additional logic for specifically setting the VectorT*
ISA is required for other scenarios, or if incorporating similar logic
to `EEJitManager::SetCpuInfo()` into `Compiler::compCompile()` is
required.

Adds new JIT debugging helpers `dIsa()` and `dIsaFlags()` to dump out
string representations of `CORINFO_InstructionSet` and
`CORINFO_InstructionSetFlags`.

Fixes #113869

* Only copy over VectorT ISAs for matching VM case

That is, AltJit on same architecture (matching VM) with RunAltJitCode=0.