Do not generate unconstructed EEType for arrays that have a template (#113951)
When `CanonFormTypeMayExist` was written, it was okay for arrays to return false since arrays were always compared structurally (it was allowed to have multiple `MethodTable`s for a single array type).

We stopped allowing this some time ago and we now have to maintain the same invariant as for any other `MethodTable` - the address of the `MethodTable` is the unique identity. We cannot allow the
compiler to generate unconstructed `MethodTable` for something that can be template-constructed at runtime.