Fix declared type of custom comparer for MaxBy and MinBy (#113944)
* Fix declared type of custom comparer for MaxBy and MinBy

Use the `TKey` output of the keySelector (not the input `TSource`) for the the custom `IComparer` since that's what the keySelector will emit.

Fixes #113878

This is a semi-**BREAKING CHANGE** but is justifiable in that the only code that could have worked before was when `TKey` and `TSource` were the same type (e.g. both `int` as in the test cases)... and that code will continue to compile correctly.

* Update to keep the old MinBy and MaxBy around as Obsolete

Ensures the binary and source compatibility in case anyone was using the broken overloads.

* PR Feedback: Fixed missing markdown escaping for < and >