Don't use `Delegate.Method` in ServiceDescriptor
The problem with this API is that in a trimmed app, accessing `Method` on a delegate type forces trimming to consider all delegate targets also targets of reflection (since we're grabbing a `MethodInfo` for whatever is the target). We do try to optimize this and only consider targets of the same delegate type, but in this case it means `Func<T, U>` and `Func<T,U,V>` which are very popular delegate types. Avoiding `Delegate.Method` is a 0.8% size saving for the Microsoft Store app.

The differences are:
* Originally this was calling `ToString` on the `MethodInfo`, which returns the name and signature of the method. I'm replacing this with `FullName` of the owning type and the name of the method. This is both more information and less information - we get owning type, we lose signature. It feels more useful nevertheless.
* When the app is trimmed and `StackTraceSupport` feature switch is set to false, this will return question marks.

The 0.8% size savings is present no matter the value of the `StackTraceSupport` switch.