Fix Unable to load shared library 'jitinterface_x64' build error on linux/mac
When crossgen is first built, it writes the executable and shared lib artifacts to
`artifacts/bin/coreclr/<os>.<arch>.<config>/<arch>/crossgen2`. On subsequent builds,
the native libs are cleaned and never copied back even though they are included as
<Content> items in crossgen.props. This can cause the build to fail later on down the
line when we attempt to use the in build crossgen on different assemblies as it won't find
the native libs it depends on.

To work around the issue, this change adds a target that runs after publish and copies
the content items to $(PublishDir). It's not clear if this is a bug in the SDK or something
we're not doing correctly. More investigation is suggested.

Fixes https://github.com/dotnet/runtime/issues/111731