Move the frames popping to 2nd pass
The SOS tests that examine crash dump of an unhandled exception were
failing with this change due to the fact that the explicit frames and
ExInfos were popped before the RaiseException. The Watson / crash dump
generation kicks in at the end of the first pass of the RaiseException,
so the frames need to be popped in the 2nd pass. I have changed it so
that they are popped at the first call to the ProcessCLRException on the
2nd pass.