Add support for ThreadStatic fields
Special static fields can have a bunch of accessors. This commit handles a few accessor types for thread static fields. In order to obtain the address of such a field, we emit an interpreter opcode that can call a helper of a certain signature, with the slot containing the helper code pointer and argument embedded in the code stream. It seems like the helpers can either point to native code or to jit compiled code.