Fix EH and stack walking on linux-x86 (#113972)
* Fix EH and stack walking on linux-x86

On x86 we don't have mechnism for unwinding of native code, so we have
to handle IL_Throw/IL_Rethrow initial context for new exception handling
differently.

Stack walking on x86 is also significantly different because we don't
produce unwind codes and rely solely on GC info produced by JIT. In
order to unwind this efficiently we always need EECodeInfo. The previous
implementation used a custom reimplementation of RtlVirtualUnwind method
which meant that we did the lookup multiple times. Another problem was
that it inconsistently treated the the "caller SP" and "return address
location". On x86 we have calling conventions that use "ret N" opcode
which pops the return adress from stack and then pops an additional
number of bytes from the stack pointer to skip over pushed arguments.
This distinction is important and most of the runtime code expects
unwound SP to point to the caller SP (ie. as if the RET N instruction
was executed). An exception is the return address hijacking which needs
the return address location instead. We now treat the SP and PCTAddr
fields of REGDISPLAY the same as on non-funclet runtime to make this
distinction respected in a consistent way.

* Update src/coreclr/vm/i386/cgenx86.cpp

* Address PR feedback

* Update src/coreclr/vm/exceptionhandling.cpp