JIT: Remove `BBF_FUNCLET_BEG` (#113975)
BBF_FUNCLET_BEG is set for the first block in each filter/handler region when we have funclets; this state can already be checked with Compiler::bbIsHandlerBeg. This PR replaces BBF_FUNCLET_BEG with a new helper method Compiler::bbIsFuncletBeg, which is identical in behavior to bbIsHandlerBeg when we have funclets, and (what should be) a const-evaluated return false when we don't use funclets at all.

This was a no-diff change locally, though it does enable compaction of the first block in a handler region -- I don't see why such a transformation should be illegal.