Stackwalk support for interpreted frames (#113900)
* Stackwalk support for interpreted frames

This change adds support for walking stack with interpreter frames
present. The StackFrameIterator now considers the interpreter frames to
be frameless frames like regular JITted / crossgened folder.
This is a basis for GC, EH and debugger stack walking. The SOS
`clrstack` and `clrstack -i` both display correct stack trace with
interpreter frames on the stack with this change.

I have added some extra handling of the case when the interpreter frames
are on top of the stack so that the clrstack can dump them.

* Fix VirtualUnwindInterpreterCallFrame

In case we were generating the caller context, the SP and ContextFlags
were missing.

* Move to single InterpreterFrame and few fixes

* Fix Unix build break

* Prevent missing override errors on Unix build

* Improve top frame lookup

* Reflect PR feedback

* Change the GET_CALLER_SP to assert and return NULL
* Remove extra assert at the call to GET_CALLER_SP
* Revert some changes removing "virtual" keyword that were not
  intentional

* Few missing virtual keywords reverts

* Revert the GetCallerSP changes

* Fix MUSL build break

* Add comment on top frame seeking