Propagate `Type.GetInterfaces` through dataflow analysis
Fixes dotnet/linker#1731

We currently maintain two invariants:

1. The types returned by the API will have `.Interfaces` annotations at minimum
2. If the parent type was annotated `.All`, the types returned by the API will also be `.All`

We do this in the logic that keeps things, but we don't do this in terms of dataflow analysis warnings (the types retrieved from the array are not annotated as such, as far as the analysis is concerned). Because of the lacking annotation, we have warning suppressions in multiple places within the framework.

This fixes it.

Opening as a draft because I'm having trouble making `foreach` of arrays work with the Roslyn analyzer. Roslyn models it as `IEnumerable` instead of array indexing (which is how foreach is actually expanded for arrays). I hope we can somehow force Roslyn to surface this as array indexing. In the worst case we'll need to live with the wart that `foreach` doesn't work, only `for`.