Add field access support to CoreCLR Interpreter (#114049)
* Add support for newobj

The newobj instruction receives the set of arguments and it needs to create the object in question, add it as the `this` argument to the constructor and return it as well. All this work is done by a single interpreter newobj instruction. Once handling the creation of the new object, the already existing call path is reused for executing the constructor. The newobj instruction thinks it receives the `this` pointer as an argument, so we can reuse the offset allocator to generate the exact offsets for each argument. We dummy initialize `this` via an intermediary opcode, so the offset allocator handles this pattern cleanly.

* Intrinsify Object ctor

We currently have limited control over what is interpreted and what is jitted. When creating a new object in the interpreter we have the problem of having to call to object ctor which is jitted. A simple work around, at least for now, is to ignore this call.

* Add support for fields

Loading and storing into instance and static fields, as well as loading the address of fields. For instance fields, this is achieved via ldind/stind opcodes, that have the offset embedded into the instruction. For static fields, we first compute the field address and then we reuse the same ldind/stind opcodes for the field access.

When loading a field from a valuetype from the interpreter stack, we use a separate opcode INTOP_MOV_SRC_OFF, which will be emitted as a normal move (with already existing opcodes), but the source offset that it operates on can point inside a VT var.

* Add support for ldloca

* Add support for ThreadStatic fields

Special static fields can have a bunch of accessors. This commit handles a few accessor types for thread static fields. In order to obtain the address of such a field, we emit an interpreter opcode that can call a helper of a certain signature, with the slot containing the helper code pointer and argument embedded in the code stream. It seems like the helpers can either point to native code or to jit compiled code.

* Zero IL locals at method entry, if initlocals flag is set

* Add support for ldind/stdind instructions

Reuse the instructions used for loading/storing into fields

* Emit memory barriers for volatile stores/writes

For simplicity we emit full memory barriers each time

* Add some more tests related to field access

Some of them are allocating objects so we disable them currently just to be safe on CI.