Fix Unable to load shared library 'jitinterface_x64' build error on linux/mac (#114012)
When crossgen is first built, it writes the executable and shared lib artifacts to `artifacts/bin/coreclr/<os>.<arch>.<config>/<arch>/crossgen2`. On subsequent builds the native libs are cleaned and never copied back (by the SDK) even though they are included as <Content> items in crossgen.props. This can cause the build to fail later on down the line when we attempt to use the in build crossgen on different assemblies as it won't find the native libs it depends on.

We need to set `ExcludeFromSingleFile=true` so that the publishing logic knows that we want to not include these files in the singlefile bundle.

See https://github.com/dotnet/designs/blob/main/accepted/2020/single-file/design_3_0.md

Fixes https://github.com/dotnet/runtime/issues/111731