Disable, by default, CLRConfig settings on DAC builds.
The majority of the CLRConfig settings should have no impact
on the code in the DAC. During investigation 4 where found
but only 2 were still used. This PR removes the 2 unused and
creates a new flag that that indicates the setting is valid for
the DAC. All others fallback to their default value.