[mono] Tweaks to bridge testing (#114203)
* Tweak disabled test for future investigation

Seems to pass with 50000 but fails with 100000.

* Run bridge tests also in tarjan only mode

Initially, we were running the bridge tests as comparison between tarjan and new bridge output. In order for the output to match, certain optimizations need to be disabled for the tarjan bridge and this configuration would be different from standard production configuration. Address this by also running the tests in tarjan only mode. While we don't verify the correctness of the generated SCCs in this configuration, we could still catch certain assertions in the implementation. For example, FauxHeavyNodeWithCycles would lead to crashes in tarjan only configuration.