[clr-interp] Add support for DOTNET_InterpHalt
This flag adds a `INTOP_BREAKPOINT` instruction at the beginning of the method. This opcode calls the `InterpBreakpoint` method. In order to make use of this debug option, you need to add a breakpoint to this method.

An alternative would be to use special instructions to trigger a trap in the debugger, but there is no clean cross platform way to do this currently. Using `__builtin_debugtrap()` on arm64 doesn't allow for step/continue, you would need to set the PC register manually to the next instruction in order to exit the trap, making it a pain to use.