Extend static constructor interpreter to support vtable-like structures (#114354)
This makes it possible to identify `struct`s that look like vtables (all fields are `delegate*` and some extras) and represents them differently within the interpreter. The different representation allows the interpreter to capture relocs within the data structure (the normal byte array approach wouldn't be able to represent these since the numerical value of a function pointer is only known at runtime).

I had to extend things so that we can allow byrefs to pass through as nint in places. Hopefully I found all the places where we would try to cast things to `ValueTypeValue` within the interpreter and we now allow `nint`-typed stack slots to also hold a byref.